+++ src/components/InstancesList/InstancesList.js
@@ -184,6 +184,8 @@ class InstancesList extends React.Component {
     const {
       precedingTitles,
       succeedingTitles,
+      childInstances,
+      parentInstances,
     } = instance;
     let copiedInstance = omit(instance, ['id', 'hrid']);
 
@@ -193,6 +195,14 @@ class InstancesList extends React.Component {
 
     if (succeedingTitles?.length) {
       copiedInstance.succeedingTitles = omitFromArray(succeedingTitles, 'id');
+    }
+
+    if (childInstances?.length) {
+      copiedInstance.childInstances = omitFromArray(childInstances, 'id');
+    }
+
+    if (parentInstances?.length) {
+      copiedInstance.parentInstances = omitFromArray(parentInstances, 'id');
     }
 
     copiedInstance = set(copiedInstance, 'source', 'FOLIO');