+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Keyboard cannot move item to other holding by drag 'n' drop. Refs UIIN-1490.
 * Sort statistical codes by code-type, then code, then name. Refs UIIN-1550.
 * Populate a value in Source field when View/Edit Inventory Holdings Record. Refs UIIN-1548.
+* Fix instance duplication when child or parent records are present. Fixes UIIN-1562.
 
 ## [7.1.2](https://github.com/folio-org/ui-inventory/tree/v7.1.2) (2021-07-14)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v7.1.1...v7.1.2)